Hello world #034.
<details><summary>Here's how it works</summary>
<p>

The string at the top is base3-encoded ascii characters, i.e. the world emojis have the values 0,1,2.

This means we can fit each character in "Hello, World" into 5 symbols.

The for-loop is essentially a big-endian base-3 decoder.

Emojis are 2 chars each, which is why we take the characters two at a time.

I didn't do a super good job of getting the loop start/end conditions right, so there's a space at the end of the input (to get an extra iteration, to trigger the code that flushes 'thing' into 'thang'), and we have to slice off the first character in `thang` at the end, which is '\0'. This is especially evil because you can't actually _see_ \0 when it's printed to console 😅

</p>
</details>